requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
psycopg2-binary
requests
jinja2
twilio
